Split up functions from makeInventory
