+++ src/lib/dataManip/makeInventory.js
@@ -1,8 +1,6 @@
 //todo: cleanup for lib-centric approach
 const path = require('path')
 
-const { getWorkspacePath } = require('../pathHelpers')
-
 const inventoryGcp = require('./inventoryGcp')
 const inventoryAws = require('./inventoryAws')
 const inventoryDo = require('./inventoryDo')
@@ -11,8 +9,7 @@ const configGcp = require('./configGcp')
 const configAws = require('./configAws')
 const configDo = require('./configDo')
 
-const binary = require('./binary')
-const envPython = require('./envPython')
+const structure = require('./structure')
 
 // const boolToString = require('./preprocessors/boolToString')
 // const dynamicEnvVar = require('./preprocessors/dynamicEnvVar')
@@ -57,25 +54,26 @@ const buildDynamicInventory = async gantreeConfigObj => {
   resolveNodeNames(gantreeConfigObj)
 
   // object to return from function
-  const o = {
-    _meta: {
-      hostvars: {
-        localhost: {
-          infra: []
-        }
-      }
-    },
-    local: {
-      hosts: ['localhost'],
-      vars: {
-        ansible_python_interpreter: await envPython.getInterpreterPath(),
-        ansible_connection: 'local'
-      }
-    },
-    all: {
-      vars: await getSharedVars({ gantreeConfigObj })
-    }
-  }
+  // const o = {
+  //   _meta: {
+  //     hostvars: {
+  //       localhost: {
+  //         infra: []
+  //       }
+  //     }
+  //   },
+  //   local: {
+  //     hosts: ['localhost'],
+  //     vars: {
+  //       ansible_python_interpreter: await envPython.getInterpreterPath(),
+  //       ansible_connection: 'local'
+  //     }
+  //   },
+  //   all: {
+  //     vars: await getSharedVars({ gantreeConfigObj })
+  //   }
+  // }
+  const o = await structure.skeleton.generate(gantreeConfigObj)
 
   const validator_list = []
 
@@ -116,76 +114,6 @@ const resolveNodeNames = gantreeConfigObj => {
     // TODO: FIXME: believe this is unused and thus obsolete
     item.infra_name = 'gantree-infra-' + item.name
   })
-}
-
-const getSharedVars = async ({ gantreeConfigObj: c }) => {
-  const ansibleGantreeVars = {
-    // ansible/gantree vars
-    gantree_root: '../',
-    gantree_control_working: getWorkspacePath(c.metadata.project, 'operation'),
-    ansible_ssh_common_args:
-      '-o StrictHostKeyChecking=no -o ControlMaster=no -o UserKnownHostsFile=/dev/null -o ServerAliveInterval=30 -o ControlPersist=60s'
-  }
-
-  const miscSharedVars = {
-    // shared vars
-    substrate_network_id: 'local_testnet', // TODO: this probably shouldn't be hard-coded
-    project_name: c.metadata.project
-  }
-
-  const binaryInvKeys = binary.resolveInvKeys(c.binary)
-
-  // console.log(binaryInvKeys)
-  // console.log("exit early")
-  // process.exit(1)
-
-  // const binaryVars = {
-  //   // required
-  //   // substrate_bin_name: binKeys.filename,
-
-  //   // optional
-  //   // substrate_binary_sha256: (binKeys.fetch && binKeys.fetch.sha256) || 'false', // TODO: not yet implemented
-
-  //   // substrate_binary_url: (binKeys.fetch && binKeys.fetch.url) || 'false',
-  //   // substrate_use_default_spec: binKeys.useBinChainSpec || 'false',
-  //   // substrate_chain_argument: binKeys.chain || 'false',
-
-  //   // substrate_binary_path: (binKeys.local && binKeys.local.path) || 'false', // TODO: not yet implemented
-
-  //   // substrate_repository_url:
-  //   //   (binKeys.repository && binKeys.repository.url) || 'false',
-  //   // substrate_local_compile:
-  //   //   (binKeys.repository && binKeys.repository.localCompile) || 'false',
-
-  //   // substrate_bootnode_argument: binKeys.bootnodes || []
-  // }
-
-  const telemetryVars = {
-    telemetry: {
-      repository: 'https://github.com/flex-dapps/substrate-telemetry.git',
-      binary_url:
-        'https://nyc3.digitaloceanspaces.com/gantree-rozifus-00/flexdapps-telemetry-0.1.0',
-      binary_name: 'telemetry',
-      src_folder: 'telemetry_src',
-      src_subfolder: 'backend',
-      operation: 'fetch'
-    },
-    substrate_telemetry_argument: c.telemetry || 'ws://127.0.0.1:8000/submit'
-  }
-
-  // console.log("----BINARY VARS----")
-  // console.log(binaryVars)
-  // console.log("EXITING EARLY")
-  // process.exit(-1)
-
-  const sharedVars = {
-    ...ansibleGantreeVars,
-    ...miscSharedVars,
-    ...binaryInvKeys,
-    ...telemetryVars
-  }
-
-  return sharedVars
 }
 
 const getNodeVars = ({ item, infra }) => {
