+++ src/lib/dataManip/structure/index.js
@@ -1,0 +1,7 @@
+const skeleton = require('./skeleton')
+
+//
+
+module.exports = {
+  skeleton
+}
